+++ CaptchaSharp/Services/AntiCaptcha/Responses/Solutions/ITaskSolution.cs
@@ -4,6 +4,6 @@ namespace CaptchaSharp.Services.AntiCaptcha.Responses.Solutions
 {
     internal interface ITaskSolution
     {
-        CaptchaResponse ToCaptchaResponse(int id);
+        CaptchaResponse ToCaptchaResponse(long id);
     }
 }
